requirements.txt
django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-gis
psycopg2-binary==2.9.3